Fixed Serial_::begin(..) definition to match arduino::HardwareSerial
This is usually not a problem becuase we use the arm-none-eabi-gcc compiler
thas has type `uint32_t` aliased to `unsigned long`. Anyway this is not
mandatory in general, and the compiler may choose to alias `uint32_t` with
`unsigned int` as it happens with llvm/clang.

Since we use clangd as language server on the Arduino IDE 2.0 the
Serial_ object and all the derived instances (Serial, SerialUSB, ...) are
not available in code-completion suggestions.

+++ cores/arduino/USB/CDC.cpp
@@ -196,12 +196,12 @@ void Serial_::enableInterrupt() {
 	usbd.epBank0EnableTransferComplete(CDC_ENDPOINT_OUT);
 }
 
-void Serial_::begin(uint32_t /* baud_count */)
+void Serial_::begin(unsigned long /* baudrate */)
 {
 	// uart config is ignored in USB-CDC
 }
 
-void Serial_::begin(uint32_t /* baud_count */, uint16_t /* config */)
+void Serial_::begin(unsigned long /* baudrate */, uint16_t /* config */)
 {
 	// uart config is ignored in USB-CDC
 }
